fix logic

+++ src/db/models/envelope.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy import (
     Boolean,
     ForeignKey,
@@ -12,7 +14,7 @@ from .base import Base
 class Envelope(Base):
     __tablename__ = "envelopes"
     name: Mapped[str] = mapped_column(String)
-    balance: Mapped[float] = mapped_column(Numeric, default=0)
+    balance: Mapped[Decimal] = mapped_column(Numeric, default=0)
     owner_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"))
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     is_savings: Mapped[bool] = mapped_column(Boolean, default=False)
