+++ src/db/models/goal.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy import (
     ForeignKey,
     Numeric,
@@ -11,7 +13,7 @@ from .base import Base
 class Goal(Base):
     __tablename__ = "goals"
     name: Mapped[str] = mapped_column(String)
-    target_amount: Mapped[float] = mapped_column(Numeric)
+    target_amount: Mapped[Decimal] = mapped_column(Numeric)
     linked_envelope_id: Mapped[int] = mapped_column(ForeignKey("envelopes.id"))
     status: Mapped[str] = mapped_column(String, default="active")  # active, archived
     phase_id: Mapped[int] = mapped_column(ForeignKey("phases.id"))
