+++ src/db/models/transaction.py
@@ -1,4 +1,5 @@
 import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     Date,
@@ -17,7 +18,7 @@ class Transaction(Base):
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     category_id: Mapped[int] = mapped_column(ForeignKey("categories.id"))
     envelope_id: Mapped[int] = mapped_column(ForeignKey("envelopes.id"))
-    amount: Mapped[float] = mapped_column(Numeric)
+    amount: Mapped[Decimal] = mapped_column(Numeric)
     transaction_date: Mapped[datetime.date] = mapped_column(Date)
     comment: Mapped[str | None]
     created_at: Mapped[datetime.datetime] = mapped_column(DateTime, server_default=func.now())
